refactor: move StudyProgress Firestore write to entities/study-progress/api (#898)
Move the StudyProgress Firestore write implementation from src/features/study/api/editStudyProgress.ts to src/entities/study-progress/api/firestore.ts.

- Export editStudyProgress from src/entities/study-progress.
- Update useEditStudyProgress to import editStudyProgress from @/entities/study-progress.
- Colocate unit tests in src/entities/study-progress/api/firestore.spec.ts.
- Mock @/shared/firebase in study feature unit tests to align with entity API loading.
- Remove src/features/study/api/editStudyProgress.ts.
