+++ src/entities/study-progress/api/firestore.ts
@@ -1,9 +1,10 @@
-import { editStudyProgressSchema, type EditStudyProgressInput } from "@/entities/study-progress";
+import type { EditStudyProgressInput } from "../model/types";
 
 import { doc, updateDoc } from "firebase/firestore";
 
 import { db } from "@/shared/firebase";
 import { getTimestamp, omitUndefined } from "@/shared/firestore";
+import { editStudyProgressSchema } from "../model/schema";
 
 export const editStudyProgress = async (uid: string, progress: EditStudyProgressInput["progress"]): Promise<void> => {
   const input = editStudyProgressSchema.parse({ uid, progress });
