+++ src/features/study/hooks/useStudyCards.spec.tsx
@@ -1,6 +1,8 @@
 import { act, renderHook } from "@testing-library/react";
 import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
 
+vi.mock("@/shared/firebase", () => ({ auth: {}, db: {} }));
+
 import { createCard, createPreferences, createDeck } from "@/test/factories";
 
 import { useStudyCards } from "./useStudyCards";
