+++ src/features/study/model/swipe.spec.ts
@@ -1,6 +1,8 @@
 import type { SwipeState } from "@/entities/preferences";
 
-import { describe, expect, it } from "vitest";
+import { describe, expect, it, vi } from "vitest";
+
+vi.mock("@/shared/firebase", () => ({ auth: {}, db: {} }));
 
 import type { StudyCard } from "./studyCard";
 import { buildStudyPatch, resolveSwipeAction } from "./swipe";
